requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
banal==1.0.6
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
easyocr==1.7.1
executing==2.0.1
fastapi==0.104.1
fastapi-sessions==0.3.2
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.44.0
fsspec==2023.10.0
greenlet==3.0.1
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httptools==0.6.1
humanfriendly==10.0
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
imutils==0.5.4
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
langdetect==1.0.9
lazy_loader==0.3
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mplcursors==0.5.2
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
onnx==1.15.0
onnxruntime==1.16.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.0.1
platformdirs==4.0.0
prompt-toolkit==3.0.40
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydyf==0.8.0
Pygments==2.16.1
pyparsing==3.1.1
pypdfium2==4.24.0
pyphen==0.14.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.5.2
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
thop==0.1.1.post2209072238
tifffile==2023.9.26
tinycss2==1.2.1
torch==2.1.0
torchaudio==2.1.0
torchdata==0.7.0
torchtext==0.16.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.210
Unidecode==1.3.7
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.9
weasyprint==60.1
webencodings==0.5.1
websockets==12.0
zipp==3.17.0
zopfli==0.2.3
